chore(chaosengine): post-PhaseB sweep — drop stale GOPRIVATE, tag k8s tests, fix stale assertion
- Remove GOPRIVATE env entries that only referenced the now-absorbed
  chaos-experiment module (notify/gateway/configcenter/blob charts +
  top-level deployment.yaml).
- Move TestGetLabel + TestCRDClient1 to a sibling integration_test.go
  guarded by `//go:build integration` so default `go test ./...` is green;
  unrelated passing tests stay in the original file.
- Update TestGetAppLabelByIndexUsesInjectableOrdering expected value from
  "svc-b" to "svc-a" to match the broadened allowlist behavior (upstream
  dc831734 included non-network-pair services in the ordering).
