+++ aegislab/helm/charts/blob/values.yaml
@@ -43,7 +43,6 @@ configMap:
 # Literal env-var values injected into the container.
 env:
   ENV: "prod"
-  GOPRIVATE: "github.com/OperationsPAI/chaos-experiment"
 
 # SSO client-side wiring.
 ssoClient:
